Use relative link for pre-commit. Upgrade test.

+++ tests/install.sh
@@ -7,6 +7,24 @@ set -e
 set -u
 export PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin
 
+usage() {
+  echo "Usage: $0 [-k] [-c DIR]"
+  echo "  -c DIR - Composer dir."
+  echo "  -k     - Keep the test."
+  exit 1
+}
+
+composer_dir=''
+keep=0
+while getopts 'kc:' c
+do
+  case "$c" in
+    c) composer_dir="$OPTARG" ;;
+    k) keep=1 ;;
+    *) usage ;;
+  esac
+done
+
 if ! command -v jq >/dev/null 2>&1; then
   echo "jq not found."
   echo "Mac: brew install jq"
@@ -15,15 +33,25 @@ if ! command -v jq >/dev/null 2>&1; then
   exit 1
 fi
 
-trap "rm -rf /tmp/code-quality-test-install.$$.*" EXIT
+if [ "$keep" -eq 0 ]; then
+  trap "rm -rf /tmp/code-quality-test-install.$$.*" EXIT
+fi
 
-install_dir=$(mktemp -d /tmp/code-quality-test-install.$$.XXXXXX)
+git_dir=$(mktemp -d /tmp/code-quality-test-install.$$.XXXXXX)
 
 package_dir=$(dirname "${BASH_SOURCE[0]}")/..
 
-cd "$install_dir"
+if [ -z "$composer_dir" ]; then
+  composer_dir="$git_dir"
+else
+  composer_dir="$git_dir/$composer_dir"
+  mkdir -pv "$composer_dir"
+fi
+
+cd "$git_dir"
 git init
 
+cd "$composer_dir"
 composer init \
   --no-interaction \
   --name='wunderio/code-quality-test' \
@@ -33,8 +61,8 @@ composer init \
   --repository='{ "type": "vcs", "url": "git@github.com:wunderio/code-quality.git" }' \
   --require-dev='wunderio/code-quality:dev-master' \
 
-composer_json="$install_dir/composer.json"
-composer_json_tmp="$install_dir/composer.json.tmp"
+composer_json="$composer_dir/composer.json"
+composer_json_tmp="$composer_dir/composer.json.tmp"
 jq '.scripts += {
   "post-install-cmd": [ "./vendor/wunderio/code-quality/install-update.sh" ],
   "post-update-cmd": [ "./vendor/wunderio/code-quality/install-update.sh" ]
@@ -45,5 +73,10 @@ mv "$composer_json_tmp" composer.json
 composer install
 composer update
 
-cat composer.json
-ls -la . .git/hooks vendor/wunderio/code-quality
+cat "$composer_dir/composer.json"
+ls -la \
+  "$git_dir" \
+  "$git_dir/.git/hooks" \
+  "$composer_dir/vendor/wunderio/code-quality"
+
+test -f "$git_dir/.git/hooks/pre-commit"
